+++ simple/src/main/java/com/ybao/pullrefreshview/simple/view/ExpandHeaderView.java
@@ -81,7 +81,7 @@ public class ExpandHeaderView extends BaseHeaderView {
     }
 
     @Override
-    public int getSpanHeight() {
+    public float getSpanHeight() {
         return loadBox.getHeight();
     }
