+++ simple/src/main/java/com/ybao/pullrefreshview/simple/view/NormalFooterView.java
@@ -85,7 +85,7 @@ public class NormalFooterView extends BaseFooterView {
     }
 
     @Override
-    public int getSpanHeight() {
+    public float getSpanHeight() {
         return getHeight();
     }
